updated image update in update profile api

+++ app/Http/Controllers/Api/ProfileController.php
@@ -34,7 +34,9 @@ class ProfileController extends Controller
             $extension = $file->getClientOriginalExtension();
             $filename_with_ext = time() . '.' . $extension;
             if ($user->image) {
-                Storage::delete('public/organization/' . $user->image);
+                if (file_exists('public/organization/' . $user->image)) {
+                    Storage::delete('public/organization/' . $user->image);
+                }
             }
             $request->file('image')->storeAs($path, $filename_with_ext);
             $user->image = $filename_with_ext;
